requirements.txt
absl-py==1.0.0
altgraph==0.17.2
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
flatbuffers==2.0
fonttools==4.31.1
future==0.18.2
gast==0.5.2
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.4.0
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.9.2
ipython==8.1.1
jedi==0.18.0
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.0
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.1
matplotlib-inline==0.1.3
mediapipe==0.8.9.1
nest-asyncio==1.5.4
numpy==1.22.3
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.0
pandas==1.4.1
parso==0.8.3
pefile==2021.9.3
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pyinstaller-hooks-contrib==2022.4
pyparsing==3.0.7
python-dateutil==2.8.1
pytz==2022.1
pywin32==303
pywin32-ctypes==0.2.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.63.0
traitlets==5.1.0
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.14.0
zipp==3.7.0
